Move ASSERT.h to smv_assert.h
ASSERT.h conflicts with assert.h, which is defined in the C standard, on case-insensitive platforms such as Windows.
